% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter}
\alias{filter}
\title{Filter entries in data.frame}
\usage{
filter(.data, ...)
}
\arguments{
\item{.data}{data.frame}

\item{...}{List of variables or name-value pairs of summary/modifications
functions.}
}
\value{
A data.table
}
\description{
Analogous function for \code{filter} in \pkg{dplyr}.
}
\details{
Currently data.table is not able to delete rows by reference,
}
\examples{
iris = as.data.table(iris)
iris \%>\% filter(Sepal.Length > 7)
iris \%>\% filter(Sepal.Length > 7,Sepal.Width > 3)
iris \%>\% filter(Sepal.Length > 7 & Sepal.Width > 3)
iris \%>\% filter(Sepal.Length == max(Sepal.Length))
}
\references{
\url{https://github.com/Rdatatable/data.table/issues/635}

\url{https://stackoverflow.com/questions/10790204/how-to-delete-a-row-by-reference-in-data-table}
}
\seealso{
\code{\link[dplyr]{filter}}
}
