% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{pull}
\alias{pull}
\title{Pull out a single variable}
\usage{
pull(.data, col)
}
\arguments{
\item{.data}{data.frame}

\item{col}{A name of column or index (should be positive).}
}
\value{
A vector
}
\description{
Analogous function for \code{pull} in \pkg{dplyr}
}
\examples{
mtcars \%>\% pull(2)
mtcars \%>\% pull(cyl)
mtcars \%>\% pull("cyl")
}
\seealso{
\code{\link[dplyr]{pull}}
}
