% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv.R
\name{read_csv}
\alias{read_csv}
\title{Convenient file reader}
\usage{
read_csv(path, utf8 = FALSE, ...)
}
\arguments{
\item{path}{File name in working directory, path to file.}

\item{utf8}{Should "UTF-8" used as the encoding? (Defaults to \code{FALSE})}

\item{...}{Other parameters passed to \code{data.table::fread}.}
}
\value{
A data.table
}
\description{
A wrapper of \code{fread} in \pkg{data.table}.
Highlighting the encoding.
}
